+++ netpyne/specs.py
@@ -327,8 +327,8 @@ class NetParams (object):
         if not label: 
             label = int(self._labelid)
             self._labelid += 1
-        secs, secLists, synMechs = utils.importCell(fileName, cellName, cellArgs, cellInstance)
-        cellRule = {'conds': conds, 'secs': secs, 'secLists': secLists}
+        secs, secLists, synMechs, globs = utils.importCell(fileName, cellName, cellArgs, cellInstance)
+        cellRule = {'conds': conds, 'secs': secs, 'secLists': secLists, 'globs': globs}
         
         # adjust cell 3d points so that soma is at location 0,0,0 
         if somaAtOrigin: