+++ netpyne/__init__.py
@@ -1,3 +1,3 @@
 
-__version__ = '0.6.6' 
+__version__ = '0.6.7' 
 __gui__ = True  # global option to enable/disable graphics