+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 - Added netParams functions: addCellParamsWeightNorm(), saveCellParamsRule() and loadCellParamsRule()
 
+- Added option to importCell based on instantiated cell (not just tempalate or class)
+
 
 # Version 0.6.6