+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - Added option to importCell based on instantiated cell (not just tempalate or class)
 
+- Fixed bug during import that set section name to 'soma' if only 1 section
 
 # Version 0.6.6